fix relations

+++ database/migrations/2022_04_30_001005_create_material_table.php
@@ -19,7 +19,7 @@ class CreateMaterialTable extends Migration
             $table->id('material_id');
             $table->string('material_idf', 50)->unique();
             $table->string('material_title', 150);
-            $table->integer('material_counter_type_id')->default(1);
+            $table->unsignedBigInteger('material_counter_type_id')->default(1);
         });
 
         Schema::table('material', function($table) {
